+++ implementations/rust/ockam/signature_bls/src/multi_public_key_vt.rs
@@ -1,6 +1,9 @@
 use crate::PublicKeyVt;
 use bls12_381_plus::{G1Affine, G1Projective};
-use core::ops::{BitOr, Not};
+use core::{
+    fmt::{self, Display},
+    ops::{BitOr, Not},
+};
 use group::Curve;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use subtle::{Choice, CtOption};
@@ -16,6 +19,12 @@ impl From<&[PublicKeyVt]> for MultiPublicKeyVt {
             g += k.0;
         }
         Self(g)
+    }
+}
+
+impl Display for MultiPublicKeyVt {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
     }
 }
 
@@ -59,7 +68,7 @@ impl MultiPublicKeyVt {
     }
 
     /// Get the byte representation of this key
-    pub fn to_bytes(&self) -> [u8; Self::BYTES] {
+    pub fn to_bytes(self) -> [u8; Self::BYTES] {
         self.0.to_affine().to_compressed()
     }