+++ implementations/rust/ockam/signature_bls/src/multi_signature.rs
@@ -1,6 +1,9 @@
 use crate::{MultiPublicKey, PublicKey, Signature};
 use bls12_381_plus::{G1Affine, G1Projective};
-use core::ops::{BitOr, Not};
+use core::{
+    fmt::{self, Display},
+    ops::{BitOr, Not},
+};
 use group::Curve;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use subtle::{Choice, CtOption};
@@ -12,6 +15,12 @@ pub struct MultiSignature(pub(crate) G1Projective);
 impl Default for MultiSignature {
     fn default() -> Self {
         Self(G1Projective::identity())
+    }
+}
+
+impl Display for MultiSignature {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
     }
 }
 
@@ -64,7 +73,7 @@ impl MultiSignature {
     }
 
     /// Get the byte sequence that represents this multisignature
-    pub fn to_bytes(&self) -> [u8; Self::BYTES] {
+    pub fn to_bytes(self) -> [u8; Self::BYTES] {
         self.0.to_affine().to_compressed()
     }