+++ implementations/rust/ockam/signature_bls/src/public_key.rs
@@ -1,6 +1,9 @@
 use super::SecretKey;
 use bls12_381_plus::{G2Affine, G2Projective};
-use core::ops::{BitOr, Not};
+use core::{
+    fmt::{self, Display},
+    ops::{BitOr, Not},
+};
 use group::Curve;
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use subtle::{Choice, CtOption};
@@ -12,6 +15,12 @@ pub struct PublicKey(pub G2Projective);
 impl Default for PublicKey {
     fn default() -> Self {
         Self(G2Projective::identity())
+    }
+}
+
+impl Display for PublicKey {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.0.fmt(f)
     }
 }
 
@@ -67,7 +76,7 @@ impl PublicKey {
     }
 
     /// Get the byte representation of this key
-    pub fn to_bytes(&self) -> [u8; Self::BYTES] {
+    pub fn to_bytes(self) -> [u8; Self::BYTES] {
         self.0.to_affine().to_compressed()
     }