+++ implementations/rust/ockam/signature_bls/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.11.0 - 2021-07-23
+### Added
+- PartialSignature
+- PartialSignatureVt
+- SecretKeyShare
+- SecretKey::split
+- SecretKey::combine
+- Signature::from_partials
+- SignatureVt::from_partials
+- Fix clippy errors
+- Implement format display for existing primitives
+
 ## v0.10.0 - 2021-07-19
 ### Changed
 - Dependencies updated.